+++ src/cli/cli.js
@@ -1,6 +1,9 @@
 import install from './commands/install.js'
 import add from './commands/add.js'
 import remove from './commands/remove.js'
+import update from './commands/update.js'
+import upgrade from './commands/upgrade.js'
+import interactive from './commands/interactive.js'
 import clean from './commands/clean.js'
 
 import debug from './options/debug.js'
@@ -13,6 +16,9 @@ const commands = [
   install,
   add,
   remove,
+  update,
+  upgrade,
+  interactive,
   clean
 ]